Add files via upload
This project focuses on analyzing a movie dataset using MapReduce and MongoDB. The dataset contains information about movie releases, including the release date and the production companies involved. The project consists of two main tasks:
Task 1: Data Preparation and AnalysisPython program using pymongo to extract the release year and the top three production companies for each movie.MapReduce program to calculate the number of movies released by the top three production companies each year.
Task 2: SortingMergeSort: MapReduce program to sort the pairs of year and company in ascending order based on the number of movies released.BucketSort: MapReduce program to sort the pairs of year and company in descending order based on the number of movies released.

The project showcases how to use MapReduce for data processing tasks and MongoDB for data storage and retrieval. It demonstrates the power and flexibility of these technologies in handling large-scale data analysis tasks efficiently.